+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 - Better claim checking for teacher page, to explain what teacher accounts are for.
 - Fixed dark mode loading.
 - Fixed #673, more visible links in paragraphs.
+- Fixed #672, more visible links in feedback with background color.
 
 ## 0.15.0 2024-01-26